restructure of postcode validation
