+++ src/test/java/uk/gov/hmcts/dts/fact/services/MapitServiceTest.java
@@ -10,10 +10,14 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.dts.fact.mapit.MapitClient;
 import uk.gov.hmcts.dts.fact.mapit.MapitData;
+import uk.gov.hmcts.dts.fact.services.validation.PostcodeValidator;
 
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
@@ -24,6 +28,9 @@ import static org.mockito.Mockito.when;
 class MapitServiceTest {
 
     private static final String RESPONSE_MESSAGE = "message";
+
+    @MockBean
+    private PostcodeValidator mockMapitValidator;
 
     @MockBean
     private MapitClient mapitClient;
@@ -126,5 +133,30 @@ class MapitServiceTest {
 
         assertThat(result).isNotPresent();
         verifyNoInteractions(mapitClient);
+    }
+
+    @Test
+    public void testValidatePostcodesSuccess() {
+        // Expect no strings to be returned if all checks have passed
+        String[] testPostcodesArray = {"M0", "MO5", "MO53", "MO533"};
+        when(mockMapitValidator.postcodeDataExists(anyString())).thenReturn(true);
+        assertEquals(0, mapitService.validatePostcodes(testPostcodesArray).length);
+    }
+
+    @Test
+    public void testValidatePostcodesInvalid() {
+        // Expect an array of strings to be returned if one or more checks have failed
+        // Note: this also tests that the stripping/trimming works, as else the mocking will not succeed
+        when(mockMapitValidator.postcodeDataExists("avalidpostcode")).thenReturn(true);
+        when(mockMapitValidator.postcodeDataExists("aninvalidpostcode")).thenReturn(false);
+        when(mockMapitValidator.postcodeDataExists("anotherinvalidpostcode")).thenReturn(false);
+
+        String[] testPostcodesArray = {"a valid postcode", "an invalid postcode", "another invalid postcode"};
+        assertArrayEquals(new String[]{testPostcodesArray[1], testPostcodesArray[2]},
+                          mapitService.validatePostcodes(testPostcodesArray));
+
+        verify(mockMapitValidator).postcodeDataExists("avalidpostcode");
+        verify(mockMapitValidator).postcodeDataExists("aninvalidpostcode");
+        verify(mockMapitValidator).postcodeDataExists("anotherinvalidpostcode");
     }
 }
